Updated for full game release
Spike dodge accuracy text is broken unless I patch a different method, so probably will never be done unless someone else does it.

Added logic to check if it's the full game or a demo to do different behavior like accounting for criticals or hiding text when judgment should be off.